# specs/008-authenticated-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Job Application Workspace Online API
  version: 0.1.0
  description: Authenticated online workspace and complete replacement operations.
servers:
  - url: https://app.example.com
security:
  - sessionCookie: []
  - desktopBearer: [workspace]
paths:
...
  /api/positions/{positionId}/application/resume:
    parameters:
      - $ref: '#/components/parameters/PositionId'
    put:
      operationId: replaceSubmittedResume
      parameters:
        - $ref: '#/components/parameters/ExpectedWorkspaceRevision'
      requestBody:
        required: true
        content:
          application/pdf:
            schema: { type: string, contentEncoding: binary }
          application/vnd.openxmlformats-officedocument.wordprocessingml.document:
            schema: { type: string, contentEncoding: binary }
      responses:
        '200': { description: Resume metadata and new workspace revision. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/RevisionConflict' }
        '413': { description: File exceeds configured size limit. }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      operationId: getSubmittedResumeAccess
      responses:
        '200':
          description: Short-lived authenticated download information.
          content:
            application/json:
              schema:
                type: object
                required: [url, expiresAt, fileName, mediaType]
                properties:
                  url: { type: string, format: uri }
                  expiresAt: { type: string, format: date-time }
                  fileName: { type: string }
                  mediaType: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    desktopBearer:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://app.example.com/api/auth/oauth2/authorize
          tokenUrl: https://app.example.com/api/auth/oauth2/token
          scopes:
            workspace: Access the owner's workspace.
  parameters:
    PositionId:
      in: path
      name: positionId
      required: true
      schema: { type: string }
    ExpectedWorkspaceRevision:
      in: header
      name: If-Match
      required: true
      schema: { type: string, description: Quoted workspace revision ETag. }
...
  responses:
    Unauthorized:
      description: Missing, expired, revoked, or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource absent or not owned by the authenticated account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RevisionConflict:
      description: Expected revision is stale; reload or create a new synchronization preview.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request or workspace package failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            issues:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string }
                  message: { type: string }
...